Moved the process and stop function to the camerahandler to allow astraction
